Extract Binding module
