+++ src/Language/Mulang/Binding.hs
@@ -1,0 +1,20 @@
+module Language.Mulang.Binding (
+  named,
+  like,
+  anyone,
+  Binding,
+  BindingPredicate) where
+
+import  Data.List (isInfixOf)
+
+type Binding = String
+type BindingPredicate = Binding -> Bool
+
+named :: String -> BindingPredicate
+named = (==)
+
+like :: String -> BindingPredicate
+like = isInfixOf
+
+anyone :: BindingPredicate
+anyone = const True
