+++ src/Language/Mulang/Inspector/Extras.hs
@@ -2,6 +2,7 @@ module Language.Mulang.Inspector.Extras (
   usesConditional,
   declaresPredicate) where
 
+import Language.Mulang.Binding
 import Language.Mulang.Explorer
 import Language.Mulang.Inspector
 import Language.Mulang.Inspector.Combiner
